feat: 대출 기능 수정

+++ isol/server-child/src/main/java/shinhan/server_child/domain/loan/repository/LoanCustomRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import shinhan.server_child.domain.loan.dto.LoanDto;
 
 public interface LoanCustomRepository {
-    List<LoanDto> findByChildID(int childId);
+    List<LoanDto> findByChildID(Long childId);
 
     void acceptLoan(int loanId);
